introduce temporal-logic subdirectory
This moves code related to temporal logics into a separate directory named
"temporal-logic".
