+++ src/ebmc/bdd_engine.cpp
@@ -12,9 +12,11 @@ Author: Daniel Kroening, daniel.kroening@inf.ethz.ch
 
 #include <ebmc/ebmc_properties.h>
 #include <ebmc/transition_system.h>
-#include <smvlang/temporal_expr.h>
 #include <solvers/bdd/miniBDD/miniBDD.h>
 #include <solvers/sat/satcheck.h>
+#include <temporal-logic/negate_property.h>
+#include <temporal-logic/temporal_expr.h>
+#include <temporal-logic/temporal_logic.h>
 #include <trans-netlist/aig_prop.h>
 #include <trans-netlist/instantiate_netlist.h>
 #include <trans-netlist/trans_to_netlist.h>
@@ -22,7 +24,6 @@ Author: Daniel Kroening, daniel.kroening@inf.ethz.ch
 #include <trans-netlist/unwind_netlist.h>
 #include <verilog/sva_expr.h>
 
-#include "negate_property.h"
 #include "report_results.h"
 
 #include <algorithm>
