+++ src/temporal-logic/Makefile
@@ -1,0 +1,16 @@
+SRC = negate_property.cpp \
+      temporal_logic.cpp \
+      #empty line
+
+include ../config.inc
+include ../common
+
+CLEANFILES = temporal-logic$(LIBEXT)
+
+all: temporal-logic$(LIBEXT)
+
+###############################################################################
+
+temporal-logic$(LIBEXT): $(OBJ)
+	$(LINKLIB)
+
