+++ src/temporal-logic/negate_property.cpp
@@ -2,7 +2,7 @@
 
 Module: Property Negation
 
-Author: Daniel Kroening, daniel.kroening@inf.ethz.ch
+Author: Daniel Kroening, dkr@amazon.com
 
 \*******************************************************************/
 
@@ -10,8 +10,9 @@ Author: Daniel Kroening, daniel.kroening@inf.ethz.ch
 
 #include <util/std_expr.h>
 
-#include <smvlang/temporal_expr.h>
 #include <verilog/sva_expr.h>
+
+#include "temporal_expr.h"
 
 /*******************************************************************\
 
@@ -27,74 +28,73 @@ Function: negate_property
 
 exprt negate_property(const exprt &expr)
 {
-  if(expr.id()==ID_and)
+  if(expr.id() == ID_and)
   {
-    exprt result=expr;
+    exprt result = expr;
     result.id(ID_or);
-    for(auto & op : result.operands())
-      op=negate_property(op);
+    for(auto &op : result.operands())
+      op = negate_property(op);
     return result;
   }
-  else if(expr.id()==ID_or)
+  else if(expr.id() == ID_or)
   {
-    exprt result=expr;
+    exprt result = expr;
     result.id(ID_and);
-    for(auto & op : result.operands())
-      op=negate_property(op);
+    for(auto &op : result.operands())
+      op = negate_property(op);
     return result;
   }
-  else if(expr.id()==ID_not)
+  else if(expr.id() == ID_not)
   {
     return to_not_expr(expr).op();
   }
-  else if(expr.id()==ID_implies ||
-          expr.id()==ID_sva_overlapped_implication)
+  else if(expr.id() == ID_implies || expr.id() == ID_sva_overlapped_implication)
   {
-    binary_exprt result=to_binary_expr(expr);
+    binary_exprt result = to_binary_expr(expr);
     result.id(ID_and);
-    result.op1()=negate_property(result.op1());
+    result.op1() = negate_property(result.op1());
     return std::move(result);
   }
-  else if(expr.id()==ID_AG)
+  else if(expr.id() == ID_AG)
   {
     unary_exprt result = to_AG_expr(expr);
     result.id(ID_EF);
-    result.op()=negate_property(result.op());
+    result.op() = negate_property(result.op());
     return std::move(result);
   }
-  else if(expr.id()==ID_sva_always)
+  else if(expr.id() == ID_sva_always)
   {
     unary_exprt result = to_sva_always_expr(expr);
     result.id(ID_sva_eventually);
-    result.op()=negate_property(result.op());
+    result.op() = negate_property(result.op());
     return std::move(result);
   }
-  else if(expr.id()==ID_AF)
+  else if(expr.id() == ID_AF)
   {
-    unary_exprt result=to_unary_expr(expr);
+    unary_exprt result = to_unary_expr(expr);
     result.id(ID_EG);
-    result.op()=negate_property(result.op());
+    result.op() = negate_property(result.op());
     return std::move(result);
   }
   else if(expr.id() == ID_sva_s_eventually)
   {
     unary_exprt result = to_sva_s_eventually_expr(expr);
     result.id(ID_sva_always);
-    result.op()=negate_property(result.op());
+    result.op() = negate_property(result.op());
     return std::move(result);
   }
-  else if(expr.id()==ID_sva_non_overlapped_implication)
+  else if(expr.id() == ID_sva_non_overlapped_implication)
   {
     // rewrite using 'next'
     unary_predicate_exprt next(ID_sva_nexttime, to_binary_expr(expr).rhs());
-    binary_exprt result=to_binary_expr(expr);
-    result.op1()=negate_property(next);
+    binary_exprt result = to_binary_expr(expr);
+    result.op1() = negate_property(next);
     return std::move(result);
   }
-  else if(expr.id()==ID_sva_nexttime)
+  else if(expr.id() == ID_sva_nexttime)
   {
     unary_exprt result = to_sva_nexttime_expr(expr);
-    result.op()=negate_property(result.op());
+    result.op() = negate_property(result.op());
     return std::move(result);
   }
   else
