+++ src/trans-word-level/property.h
@@ -38,7 +38,4 @@ symbol_exprt lasso_symbol(std::size_t k, std::size_t i);
 /// Returns true iff the given property requires lasso constraints for BMC.
 bool requires_lasso_constraints(const exprt &);
 
-/// Returns true iff the given expression contains a temporal operator
-bool has_temporal_operator(const exprt &);
-
 #endif
